feat(test): 重整human目录测试文件结构
将6个测试文件重组为2个标准命名的测试文件：
- 删除孤立文件: api_test.go, complete_test.go, coverage_test.go, edge_cases_test.go
- 合并到: human_test.go (55个测试函数 + 5个基准测试)
- 保留: locale_test.go (locale特定测试)
- 维持全面测试覆盖率，遵循Go测试文件命名规范
